SMBD/requirements_minimal.txt
# Minimal requirements untuk Python 3.13
Flask
opencv-python
ultralytics
mysql-connector-python
python-dotenv
